src/avatar_animation.py
import numpy as np
from PIL import Image
from models.fomm.demo import animate

class AvatarAnimator:
    def __init__(self, avatar_image_path):
        self.avatar_image = Image.open(avatar_image_path)
    
    def animate_avatar(self, motion_data):
        # Convert motion data into driver frames
        driver_frame = self._generate_driver_frame(motion_data)
        return animate(self.avatar_image, driver_frame)

    def _generate_driver_frame(self, motion_data):
        # Map facial motion data to a driver frame
        return np.array(motion_data)  # Placeholder logic
