requirements.txt
opencv-python
mediapipe
pyvirtualcam
torch
torchvision
PyQt5
numpy
pillow
